\name{mongoConnect}
\alias{mongoConnect}
\title{Connect to mongo...}
\usage{mongoConnect(host="localhost", port=27017)}
\description{Connect to mongo}
\details{Returns a socket connection with mongo on the other end.}
\value{socket connection}
\arguments{\item{host}{the IP address where mongo is hosted}
\item{port}{the port where mongo is hosted}}
