\name{mongoCount}
\alias{mongoCount}
\title{Send a command to Mongo...}
\usage{mongoCount(attempts=10)}
\description{Send a command to Mongo}
\details{Sends a command to Mongo.
See <URL> for documentation.}
\value{an integer, a count of the collection's items.}
\arguments{\item{attempts}{the number of times to try reading a response}}
