\name{mongoGetMore}
\alias{mongoGetMore}
\title{Gets more from cursor...}
\usage{mongoGetMore(connection, query)}
\description{Gets more from cursor}
\details{Very exciting}
\value{more results}
\arguments{\item{connection}{}
\item{query}{an old query that you want to continue getting results from}}
