\name{op_kill_cursors}
\alias{op_kill_cursors}
\title{OP_KILL_CURSORS...}
\usage{op_kill_cursors(collection="test.$cmd", cursor.ids, to_return=10)}
\description{OP_KILL_CURSORS}
\details{One of the basic mongo messages.}
\value{a raw vector encoding the query}
\arguments{\item{collection}{the name of the collection to query}
\item{cursor.ids}{a list of cursor ids}}
