\name{op_query}
\alias{op_query}
\title{OP_QUERY...}
\usage{op_query(collection, doc=list(), to_return=200)}
\description{OP_QUERY}
\details{One of the basic mongo messages.}
\value{a raw vector encoding the query}
\arguments{\item{collection}{the name of the collection to query}
\item{query_list}{the query object, given as an R list}}
