\name{op_query}
\alias{op_query}
\title{OP_QUERY...}
\usage{op_query(collection="test.$cmd", query_list=list(count = "foo", query =
    list(), fields = NULL))}
\description{OP_QUERY}
\details{See <URL>
This formats a message as a query.}
\value{raw vector}
\arguments{\item{collection}{collection to query}}
